Move and rename "SDK guide/allocation-management-ra.md".
This page is the only thing under Resource Allocators -> Custom Portal
Integration -> SDG Guide. It is called "Allocation Lifecycle
Management by Resource Allocator" but only contains information about
the rather specific task of "Listing associations between users and
resources". The information is for the API, not the SDK.

This seems to be because it was copied from Service Providers -> SDK
Guide -> "Allocation Lifecycle Management by Service Provider".

Fix this by moving this file to the API Guide and renaming it to
listing-users-resources.md with title "Listing associations between
users and resources".

As the "SDK Guide" menu for Custom Portal Integration is now empty,
remove it. We still talk about the SDK in resource-allocators.md and
link to the information page under the SDK Guide for Service
Providers. That should be enough for now.